Refactored some Gradio UI helper functions out of rss_client.py
